Refactor: Consolidate classification loss functions into a single file
