+++ model/loss.py
@@ -1,3 +1,2 @@
-from .loss_functions.binary_classification import *
-from .loss_functions.multi_classification import *
+from .loss_functions.classification import *
 from .loss_functions.regression import *
